[front] refactor: move autoSuggestions.ts in a dedicated folder
